# dennic_api_gateway/api/docs/swagger.yaml
definitions:
...
  model_booking_service.CreateDoctorNotesReq:
    properties:
      appointment_id:
        type: integer
      doctor_id:
        type: string
      patient_id:
        type: string
      prescription:
        type: string
    type: object
...
  model_booking_service.DoctorNote:
    properties:
      appointment_id:
        type: integer
      created_at:
        type: string
      doctor_id:
        type: string
      id:
        type: integer
      patient_id:
        type: string
      prescription:
        type: string
      updated_at:
        type: string
    type: object
  model_booking_service.DoctorNotesType:
    properties:
      count:
        type: integer
      doctor_notes:
        items:
          $ref: '#/definitions/model_booking_service.DoctorNote'
        type: array
    type: object
...
  model_booking_service.UpdateDoctorNoteReq:
    properties:
      appointment_id:
        type: integer
      doctor_id:
        type: string
      doctor_notes_id:
        type: string
      patient_id:
        type: string
      prescription:
        type: string
    type: object
...
  model_common.ResponseError:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  model_common.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/model_common.ResponseError'
    type: object
...
  models.StatusRes:
    properties:
      status:
        type: boolean
    type: object
host: dennic.uz:9050
info:
  contact: {}
  title: Dennic Project
  version: "1.7"
paths:
...
  /v1/doctor-notes:
    delete:
      consumes:
      - application/json
      description: DeleteDoctorNote - API to delete a doctor note
      parameters:
      - in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StatusRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: DeleteDoctorNote
      tags:
      - Doctor Note
    get:
      consumes:
      - application/json
      description: ListDoctorNotes - API to list doctor notes
      parameters:
      - description: searchField
        enum:
        - prescription
        in: query
        name: searchField
        type: string
      - in: query
        name: limit
        type: string
      - in: query
        name: order_by
        type: string
      - in: query
        name: page
        type: string
      - in: query
        name: value
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.DoctorNotesType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: ListDoctorNotes
      tags:
      - Doctor Note
    post:
      consumes:
      - application/json
      description: CreateDoctorNote - Api for create doctor note
      parameters:
      - description: CreateDoctorNotesReq
        in: body
        name: CreateDoctorNotesReq
        required: true
        schema:
          $ref: '#/definitions/model_booking_service.CreateDoctorNotesReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.DoctorNote'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: CreateDoctorNote
      tags:
      - Doctor Note
    put:
      consumes:
      - application/json
      description: UpdateDoctorNote - API to update a doctor note
      parameters:
      - description: UpdateDoctorNoteReq
        in: body
        name: UpdateDoctorNoteReq
        required: true
        schema:
          $ref: '#/definitions/model_booking_service.UpdateDoctorNoteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.DoctorNote'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: UpdateDoctorNote
      tags:
      - Doctor Note
  /v1/doctor-notes/get:
    get:
      consumes:
      - application/json
      description: GetDoctorNote - API to get doctor note by ID
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.DoctorNote'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: GetDoctorNote
      tags:
      - Doctor Note
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"